One to Many Eloquent Relations
with migration
assigning relationship
querying relationship

+++ database/migrations/2020_04_06_224327_create_profiles_table.php
@@ -17,7 +17,7 @@ class CreateProfilesTable extends Migration
             $table->increments('id');
             $table->timestamps();
 
-            $table->unsignedBigInteger('author_id')->unique();
+            $table->unsignedInteger('author_id')->unique();
             $table->foreign('author_id')->references('id')->on('authors');
         });
     }
